% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster2star.R
\name{cen2star}
\alias{cen2star}
\title{Transform Cluster To Star}
\usage{
cen2star(fit = fit)
}
\arguments{
\item{The}{output object from kmeans clustering}
}
\value{
The input object plus the star rating.
}
\description{
The clusters are transformed to star from 1 to 5 according to their
mean clustering scores
}
\examples{
fit <- restar(fit)

}
