% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{look_up}
\alias{look_up}
\title{Index Look-Up}
\usage{
look_up(centroid)
}
\arguments{
\item{centroid}{The controid is the center of the corresponding clustrer.}
}
\value{
A lookup function.
}
\description{
Index Look-Up
}
