% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rating.R
\name{rating}
\alias{rating}
\alias{rating.default}
\title{S3 Generic Function For Hospital Star Rating}
\usage{
rating(x, method, score_col)

\method{rating}{default}(x, method = c("rclus", "kmeans", "na"))
}
\arguments{
\item{x}{Winsorized summary score returned from relvm2 which contains the
wisorized summary score.}

\item{method}{Kmean cluster algorithm.\itemize{ \item kmeans: kmeans
clustering. \item rclus: rapid clustering in rclus package.}}

\item{score_col}{The column name in a data frame as input for clustering.}
}
\value{
Hospital rating stars 1, 2, 3, 4, and 5. star 5 indicates the best
  hospital in USA.
}
\description{
Hospital star rating from wisorized summary scores.
}
\seealso{
\code{\link{rating.data.frame}}
}
