% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rclus.R
\name{rclus}
\alias{rclus}
\title{Rapid Clustering}
\usage{
rclus(dat, seeds = 5, maxiter = 1000)
}
\arguments{
\item{dat}{An one-dimension array data.}

\item{seeds}{The number of start seeds.}

\item{maxiter}{The maximum number of iteration specified.}
}
\value{
A clustering object including clusters, actual iteration number,
  cluster centers and the size of each cluster, etc.
}
\description{
Rapid Clustering. It works for One Dimensional array data only currently.
}
